feat: update api content and changes for 25w03a
Changes:
- Entity.moveTo is renamed snapTo
- Entities are moving slowly to components and hiding their data
  accessors, but nothing monumental yet on this front.

There will be more developments in the future to harness the usage of
the now-fully-included GameTest framework from Mojang. We'll be able to
use this to our better advantage to verify several thousand interactions
that otherwise manual testing will often forget.

See https://minecraft.wiki/w/Java_Edition_25w03a

+++ src/mixins/java/org/spongepowered/common/mixin/core/world/level/block/PointedDripstoneBlockMixin.java
@@ -45,7 +45,7 @@ public abstract class PointedDripstoneBlockMixin extends BlockMixin {
         at = @At(value = "INVOKE", target = "Lnet/minecraft/world/damagesource/DamageSources;stalagmite()Lnet/minecraft/world/damagesource/DamageSource;"))
     private DamageSource impl$spongeRedirectForFireDamage(
         final DamageSources instance, final Level level, final BlockState state,
-        final BlockPos pos,final  Entity ignored, final float distance
+        final BlockPos pos,final  Entity ignored, final double distance
     ) {
         final DamageSource source = instance.stalagmite();
         if (level.isClientSide) { // Short Circuit
